requirements.txt
pyyaml
openai 
python-dotenv 
langchain_weaviate 
langchain_community 
langchain_core 
langchain-huggingface
langchain-openai
rank-bm25
flashrank
sentence-transformers
unstructured
pymongo 
IPython 
gradio